+++ packages/api/test/integration/main.spec.js
@@ -7,7 +7,7 @@ const fixtures = require('../utils/fixtures')
 const StateTransitionEnum = require('../../src/enums/StateTransitionEnum')
 const { getKnex } = require('../../src/utils')
 const tenderdashRpc = require('../../src/tenderdashRpc')
-const DAPI = require('../../src/dapi')
+const DAPI = require('../../src/DAPI')
 
 const genesisTime = new Date(0)
 const blockDiffTime = 2 * 3600 * 1000
